% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{add_index}
\alias{add_index}
\title{Add an index to a database.}
\usage{
add_index(db_path, index_name, index_table, index_field)
}
\arguments{
\item{db_path}{The file path to the database location.}

\item{index_name}{The unique name given to the index.}

\item{index_table}{The table in which the field to be indexed is located.}

\item{index_field}{The field to be indexed.}
}
\description{
Indexing allows for query speed improvements through the generation of an index on a specified field.
This is performed singularly for each field.
}
