% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dates.R
\name{convert_dates}
\alias{convert_dates}
\title{Convert a specified field in a dataset to a date field.}
\usage{
convert_dates(data, date_fields, format)
}
\arguments{
\item{data}{the dataset in which a field must be converted.}

\item{date_fields}{string vector, the field to be converted.}

\item{format}{string, the input format of the field to be converted, default '\%d/\%m/\%Y'.}
}
\description{
Note date format provided is input format of the field to be converted.
}
