% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{db_tables}
\alias{db_tables}
\title{Display table(s) loaded in a database.}
\usage{
db_tables(db_path)
}
\arguments{
\item{db_path}{string, the file path to the database location.}
}
\description{
Display table(s) loaded in a database.
}
