% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_db.R
\name{delete_db}
\alias{delete_db}
\title{Delete an entire database.}
\usage{
delete_db(db_path)
}
\arguments{
\item{db_path}{The file path to the database location.}
}
\description{
Prompted with y/n input on deletion.
}
