% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_n.R
\name{first_n}
\alias{first_n}
\title{Display the first n rows (records)}
\usage{
first_n(db_path, table_name, n)
}
\arguments{
\item{db_path}{The file path to the database location.}

\item{table_name}{The name of the table to be interrogated.}

\item{n}{Integer of the number of rows to be displayed (default = 10)}
}
\description{
Display the first n rows (records)
}
