% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cprd.R
\name{list_cprd}
\alias{list_cprd}
\title{List all CPRD GOLD files in a specified location.}
\usage{
list_cprd(file_location, folder = FALSE, zip = TRUE)
}
\arguments{
\item{file_location}{string, the file path location of files to be listed.}

\item{folder}{logical, the use of sub-folders.
*'TRUE': Sub-folders used, the data are in sub-folders specific to a table.
*'FLASE' (default): no sub-folders are used, the data in single location.}

\item{zip}{logical, if the files are compressed.
*'TRUE' (default): files compressed.
*'FALSE': files uncompressed.}
}
\description{
List CPRD GOLD files in only a single location, can accept the use of sub-folders:
i.e.
C:/CPRD_Extract/patient_data, C:/CPRD_Extract/practice_data,
vs
C:/CPRD_Extract/Patient/patient_data, C:/CPRD_Extract/Practice/practice_data,
}
