% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List all files of a specified type in a specified location.}
\usage{
list_files(file_location, file_type)
}
\arguments{
\item{file_location}{string, the file path location of files to be listed.}

\item{file_type}{string, the extension type of the files to be listed.}
}
\description{
List all files of a specified type in a specified location.
}
