% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_additional.R
\name{load_additional}
\alias{load_additional}
\title{Load additional data into the database}
\usage{
load_additional(db_path, file_location, table_name, type, overwrite)
}
\arguments{
\item{db_path}{The file path to the database location.}

\item{file_location}{The file path location of the file to be loaded.}

\item{table_name}{The name of the table to be loaded (or already loaded) in the database.}

\item{type}{Character file type.}

\item{overwrite}{Logical indicator if tables already exist in database and so requires overwriting.}
}
\description{
Load in CPRD/non-CPRD specific files
Accommodates varied file types
\itemize{
\item file format type: .txt, .csv, .excel, .xl, .xls, .xlsx, .dta, .rds
}
}
\details{
Note non-automated, loads only singular file into a table
Note simple data import, no additional data importing specification available
If more complex data importing specification needed, use CPRDTools::load_global
}
