% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cprd.R
\name{load_cprd}
\alias{load_cprd}
\title{Load all CPRD GOLD data.}
\usage{
load_cprd(
  db_path,
  file_location,
  tables_to_load,
  folder = FALSE,
  zip = TRUE,
  load_mapping = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{db_path}{string, the file path to the database location.}

\item{file_location}{string, the file path location of files to be loaded.}

\item{folder}{logical, the use of sub-folders.
*'TRUE': Sub-folders used, the data are in sub-folders specific to a table.
*'FLASE' (default): no sub-folders are used, the data in single location.}

\item{zip}{logical, if the files are compressed.
*'TRUE' (default): files compressed.
*'FALSE': files uncompressed.}

\item{load_mapping}{Logical indicator if the mapping/look-up files are to be loaded.
*'TRUE': no lookup mapping table required.
*'FALSE' (default): load CPRD GOLD lookup mapping table.}

\item{overwrite}{logical, if table already exist in database and requires overwriting.
*'TRUE': table exists, delete table and write in new data.
*'FALSE' (default): new table to be loaded.}

\item{table_name}{string vector, the name of the table to be loaded (or already loaded) in the database.
*can specify 'all', 'All' or 'ALL' or c('patient','practice')}
}
\description{
Loads all CPRD GOLD datasets: Additional, Clinical, Consultation, Immunisation, Patient,
Practice, Referral, Staff, Test and Therapy into a SQLite database.
}
\details{
Able to automatically load in all or selected tables.
Appends multiple .txt (.zip) files together to form a CPRD GOLD tables.
Makes use the CPRDTools::list_cprd() function to define a list of all CPRD GOLD datasets available in a given location.

Note mapping is used to map CPRD GOLD description codes to a description file, the 'lookup' file.
This file is derived from CPRD
}
