% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_global.R
\name{load_global}
\alias{load_global}
\title{Load table with a file from the global environment}
\usage{
load_global(db_path, file_to_load, table_name, overwrite)
}
\arguments{
\item{db_path}{The file path to the database location.}

\item{file_to_load}{The file in the global environment to be loaded}

\item{table_name}{The name of the table to be loaded (or already loaded) in the database.}

\item{overwrite}{Logical indicator if tables already exist in database and so requires overwriting.}
}
\description{
Best used when importing or manipulating data prior to loading into database
}
