% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_direct.R
\name{query_direct}
\alias{query_direct}
\title{Execute direct SQL query code.}
\usage{
query_direct(db_path, query)
}
\arguments{
\item{db_path}{string, the file path to the database location.}

\item{query}{string, the SQL query code.}
}
\description{
Allows for all the complexity and customisability available in an SQLite SELECT statement.
}
