% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_direct.R
\name{statement_direct}
\alias{statement_direct}
\title{Execute direct SQL statement.}
\usage{
statement_direct(db_path, statement)
}
\arguments{
\item{db_path}{string, the file path to the database location.}

\item{statement}{string, the SQL statement to be submitted.}
}
\description{
Used for data manipulation, database maintenance and improvements.
i.e. UPDATE, DELETE, INSERT INTO, DROP TABLE
Note does not output data.
}
