% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_fields.R
\name{table_fields}
\alias{table_fields}
\title{Display the fields in a table.}
\usage{
table_fields(db_path, table_name)
}
\arguments{
\item{db_path}{string, the file path to the database location.}

\item{table_name}{string, the name of the table to be interrogated.}
}
\description{
Display the fields in a table.
}
