% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_table.R
\name{update_table}
\alias{update_table}
\title{Update an existing CPRD GOLD table.}
\usage{
update_table(db_path, file_location, table_name)
}
\arguments{
\item{db_path}{string, the file path to the database location.}

\item{file_location}{string, the file path location of the file to be loaded.}

\item{table_name}{string, the name of the table in the database to be updated.}
}
\description{
Appends a single CPRD GOLD specific .txt (.zip) file to an existing table.
}
