% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_table.R
\name{delete_table}
\alias{delete_table}
\title{Delete a single or multiple tables in a database.}
\usage{
delete_table(db_path, remove_tables)
}
\arguments{
\item{db_path}{string, the file path to the database location.}

\item{remove_tables}{string vector, the table(s) to be removed.}
}
\description{
Delete a single or multiple tables in a database.
}
