% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cprd.R
\name{list_cprd}
\alias{list_cprd}
\title{List all CPRD GOLD files in a specified location.}
\usage{
list_cprd(file_location, folder, zip)
}
\arguments{
\item{file_location}{The file path location of files to be listed.}

\item{folder}{Logical indicator of the use of sub-folders.}

\item{zip}{Logical indicator if the files are compressed.}
}
\value{
File location, data frame of CPRD GOLD files and data frame of CPRD GOLD tables (datasets)
}
\description{
List CPRD GOLD files in only a single location, can accept the use of sub-foldering
i.e. C:/CPRD_Extract/patient_data, C:/CPRD_Extract/practice_data,
vs C:/CPRD_Extract/Patient/patient_data, C:/CPRD_Extract/Practice/practice_data,
}
