% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_additional.R
\name{load_additional}
\alias{load_additional}
\title{Load additional data into a database.}
\usage{
load_additional(db_path, file_location, table_name, type, overwrite = FALSE)
}
\arguments{
\item{db_path}{string, the file path to the database location.}

\item{file_location}{string, the file path location of files to be loaded.}

\item{table_name}{string, the name of the table to be loaded (or already loaded) in the database.}

\item{type}{character file type: '.txt', '.csv', '.excel', '.xl', '.xls', '.xlsx', '.dta' or '.rds'}

\item{overwrite}{logical, if table already exist in database and requires overwriting.
*'TRUE': table exists, delete table and write in new data.
*'FALSE' (default): new table to be loaded.}
}
\description{
Load in CPRD/non-CPRD specific files.
Accommodates varied file types.
}
\details{
Note non-automated, loads only singular file into a table.
Note simple data import, no additional data importing specification available.
If more complex data importing specification needed, use CPRDTools::load_global.
}
