% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cprd.R
\name{load_cprd}
\alias{load_cprd}
\title{Load all CPRD GOLD data}
\usage{
load_cprd(
  db_path,
  file_location,
  tables_to_load,
  folder,
  zip,
  load_mapping,
  overwrite
)
}
\arguments{
\item{db_path}{The file path to database location.}

\item{file_location}{The file path location of files to be listed.}

\item{tables_to_load}{The CPRD GOLD tables (datasets) to be loaded.}

\item{folder}{Logical indicator of the use of sub-folders.}

\item{zip}{Logical indicator if the files are compressed.}

\item{load_mapping}{Logical indicator if the mapping files are to be loaded.}

\item{overwrite}{Logical indicator if tables already exist in database and so require their overwriting.}
}
\description{
Loads all CPRD GOLD datasets: Additional, Clinical, Consultation, Immunisation, Patient,
Practice, Referral, Staff, Test and Therapy into a SQLite database
}
