% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_table.R
\name{rename_table}
\alias{rename_table}
\title{Rename a table in a database.}
\usage{
rename_table(db_path, old_name, new_name)
}
\arguments{
\item{db_path}{string, the file path to the database location.}

\item{old_name}{string, the table to be renamed.}

\item{new_name}{string, the new table name.}
}
\description{
Rename a table in a database.
}
